Rename requirements.py to performances.py
